feat(index-cache): abstract `IndexCache` to be shared by multi types of indexes
